Module 6 Programming Assignment 1.
Task is to implement a layer 2 firewall that runs alongside the MAC learning module on the Pyretic runtime.
The firewall application is provided with a list of MAC address pairs i.e., access control list (ACLs).
When a connection establishes between the controller and the switch, the application installs static
flow rule entries in the OpenFlow table to disable all communication between each MAC pair.

pyretic_firewall.py: a sekleton class which you will update with the logic for installing firewall rules.
firewall-policies.csv:  a list of MAC pairs (i.e., policies) read as input by the firewall application.
submit.py: used to submit your code and output to the coursera servers for grading.